Refactored all JSON aspects of hostinfo messages into a new protocol/hostinfo package
