+++ hostinfo/hostinfo.go
@@ -18,7 +18,6 @@
 package hostinfo
 
 import (
-	"encoding/json"
 	"github.com/racker/rackspace-monitoring-poller/check"
 	"github.com/racker/rackspace-monitoring-poller/protocol"
 )
@@ -26,26 +25,6 @@ import (
 type HostInfo interface {
 	Run() (*check.CheckResult, error)
 	BuildResult(cr *check.CheckResult) interface{}
-}
-
-type HostInfoBase struct {
-	Type string `json:"type"`
-}
-
-func NewHostInfo(rawParams json.RawMessage) HostInfo {
-	hinfo := &HostInfoBase{}
-	err := json.Unmarshal(rawParams, &hinfo)
-	if err != nil {
-		return nil
-	}
-	switch hinfo.Type {
-	case "MEMORY":
-		return NewHostInfoMemory(hinfo)
-	case "CPU":
-		return NewHostInfoCpu(hinfo)
-	default:
-		return nil
-	}
 }
 
 func NewHostInfoResponse(cr *check.CheckResult, f *protocol.FrameMsg, hinfo HostInfo) *protocol.HostInfoResponse {
