+++ hostinfo/hostinfo_memory.go
@@ -23,32 +23,17 @@ import (
 	"github.com/racker/rackspace-monitoring-poller/metric"
 	"github.com/racker/rackspace-monitoring-poller/utils"
 	"github.com/shirou/gopsutil/mem"
+	"github.com/racker/rackspace-monitoring-poller/protocol/hostinfo"
 )
 
 type HostInfoMemory struct {
-	HostInfoBase
+	hostinfo.HostInfoBase
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 // HostInfo Memory
 
-type HostInfoMemoryMetrics struct {
-	UsedPercentage     float64 `json:"used_percentage"`
-	Free               uint64  `json:"free"`
-	Total              uint64  `json:"total"`
-	Used               uint64  `json:"used"`
-	SwapFree           uint64  `json:"swap_free"`
-	SwapTotal          uint64  `json:"swap_total"`
-	SwapUsed           uint64  `json:"swap_used"`
-	SwapUsedPercentage float64 `json:"swap_percentage"`
-}
-
-type HostInfoMemoryResult struct {
-	Metrics   HostInfoMemoryMetrics `json:"metrics"`
-	Timestamp int64                 `json:"timestamp"`
-}
-
-func NewHostInfoMemory(base *HostInfoBase) HostInfo {
+func NewHostInfoMemory(base *hostinfo.HostInfoBase) HostInfo {
 	return &HostInfoMemory{HostInfoBase: *base}
 }
 
@@ -77,7 +62,7 @@ func (*HostInfoMemory) Run() (*check.CheckResult, error) {
 }
 
 func (*HostInfoMemory) BuildResult(cr *check.CheckResult) interface{} {
-	result := &HostInfoMemoryResult{}
+	result := &hostinfo.HostInfoMemoryResult{}
 
 	result.Timestamp = utils.NowTimestampMillis()
 	result.Metrics.UsedPercentage, _ = cr.GetMetric("UsedPercentage").ToFloat64()
