+++ protocol/hostinfo/cpu.go
@@ -1,0 +1,5 @@
+package hostinfo
+
+type HostInfoCpuResult struct {
+}
+
