+++ protocol/hostinfo/hostinfo.go
@@ -1,0 +1,6 @@
+package hostinfo
+
+
+type HostInfoBase struct {
+	Type string `json:"type"`
+}
